Made relation use long type instead of int

+++ PrestaSharp/Entities/delivery.cs
@@ -7,8 +7,8 @@ namespace Bukimedia.PrestaSharp.Entities
     {
         public long? id { get; set; }
         public long id_carrier { get; set; }
-        public int id_range_price { get; set; }
-        public int id_range_weight { get; set; }
+        public long id_range_price { get; set; }
+        public long id_range_weight { get; set; }
         public long? id_zone { get; set; }
         public long? id_shop { get; set; }
         public long? id_shop_group { get; set; }
